+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 CHANGELOG
 =========
+v1.1.0 (2014-04-03)
+-------------------
+* Added check to required data
+* Set default value for format parameter 
+
 v1.0.0 (2014-04-01)
 -------------------
 * First version with unitTests and examples